Split reverse junction exports by file
