+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/types/ValidResolvedJunctionConfig.ts
@@ -1,0 +1,6 @@
+import { type ValidJunctionConfig } from '@/object-record/record-field/ui/utils/junction/types/ValidJunctionConfig';
+
+export type ValidResolvedJunctionConfig = ValidJunctionConfig & {
+  direction: 'forward' | 'reverse';
+  isValid: true;
+};
