+++ packages/twenty-front/src/modules/settings/data-model/types/SettingsDataModelFieldEditFormValues.ts
@@ -1,0 +1,8 @@
+import type { settingsFieldFormSchema } from '@/settings/data-model/fields/forms/validation-schemas/settingsFieldFormSchema';
+import { type z } from 'zod';
+
+// TODO: Remove `any` once every field-settings form uses the discriminated schema.
+export type SettingsDataModelFieldEditFormValues = z.infer<
+  ReturnType<typeof settingsFieldFormSchema>
+> &
+  any;
